+++ sidekick/cli_template/bricks/package/__brick__/lib/src/{{#lowerCase}}{{name}}{{/lowerCase}}_command_runner.dart
@@ -2,5 +2,5 @@ import 'package:args/command_runner.dart';
 
 class {{#titleCase}}{{name}}{{/titleCase}}CommandRunner extends CommandRunner {
   {{#titleCase}}{{name}}{{/titleCase}}CommandRunner()
-      : super('Sidekick', 'A convenience script that simplifies setting up and running your project project.');
+      : super('Sidekick', 'A convenience script that simplifies setting up and running your project.');
 }