+++ sidekick/lib/src/templates/entrypoint_bundle.g.dart
@@ -1,0 +1,23 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: prefer_single_quotes, public_member_api_docs, lines_longer_than_80_chars, implicit_dynamic_list_literal, implicit_dynamic_map_literal
+
+import 'package:mason/mason.dart';
+
+final entrypointBundle = MasonBundle.fromJson(<String, dynamic>{
+  "files": [
+    {
+      "path": "entrypoint.sh",
+      "data":
+          "IyEvdXNyL2Jpbi9lbnYgYmFzaApzZXQgLWUKCiMgQXR0ZW1wdCB0byBzZXQgRU5UUllQT0lOVF9IT01FCiMgUmVzb2x2ZSBsaW5rczogJDAgbWF5IGJlIGEgbGluawpQUkc9IiQwIgojIE5lZWQgdGhpcyBmb3IgcmVsYXRpdmUgc3ltbGlua3MuCndoaWxlIFsgLWggIiRQUkciIF07IGRvCiAgbHM9JChscyAtbGQgIiRQUkciKQogIGxpbms9JChleHByICIkbHMiIDogJy4qLT4gXCguKlwpJCcpCiAgaWYgZXhwciAiJGxpbmsiIDogJy8uKicgPi9kZXYvbnVsbDsgdGhlbgogICAgUFJHPSIkbGluayIKICBlbHNlCiAgICBQUkc9JChkaXJuYW1lICIkUFJHIikiLyRsaW5rIgogIGZpCmRvbmUKU0FWRUQ9IiQocHdkKSIKY2QgIiQoZGlybmFtZSAiJFBSRyIpLyIgPi9kZXYvbnVsbApleHBvcnQgU0lERUtJQ0tfRU5UUllQT0lOVF9IT01FPSIkKHB3ZCAtUCkiCmNkICIkU0FWRUQiID4vZGV2L251bGwKCiIke1NJREVLSUNLX0VOVFJZUE9JTlRfSE9NRX0ve3t7cGFja2FnZVBhdGh9fX0vdG9vbC9ydW4uc2giICIkQCI=",
+      "type": "text"
+    }
+  ],
+  "hooks": [],
+  "name": "entrypoint",
+  "description": "Creates the sidekick entrypoint",
+  "version": "1.0.0",
+  "vars": {
+    "name": {"type": "string"},
+    "packagePath": {"type": "string"}
+  }
+});